+++ .circleci/config.yml
@@ -89,11 +89,11 @@ jobs:
       - run:
           name: Migrate database
           command: |
-            pipenv run integreat-cms-cli makemigrations cms --settings=integreat_cms.backend.circleci_settings
-            pipenv run integreat-cms-cli migrate --settings=integreat_cms.backend.circleci_settings
+            pipenv run integreat-cms-cli makemigrations cms --settings=integreat_cms.core.circleci_settings
+            pipenv run integreat-cms-cli migrate --settings=integreat_cms.core.circleci_settings
       - run:
           name: Run tests
-          command: pipenv run integreat-cms-cli test integreat_cms --set=COVERAGE --settings=integreat_cms.backend.circleci_settings
+          command: pipenv run integreat-cms-cli test integreat_cms --set=COVERAGE --settings=integreat_cms.core.circleci_settings
       - store_artifacts:
           path: htmlcov
   check-translations:
@@ -122,7 +122,7 @@ jobs:
           name: Compile translation file
           command: |
             cd src/integreat_cms
-            pipenv run integreat-cms-cli compilemessages --settings=integreat_cms.backend.circleci_settings
+            pipenv run integreat-cms-cli compilemessages --settings=integreat_cms.core.circleci_settings
       - persist_to_workspace:
           root: .
           paths:
