+++ .pylintrc
@@ -46,7 +46,7 @@ limit-inference-results=100
 # usually to register additional checkers.
 load-plugins=pylint_django,pylint.extensions.docparams
 
-django-settings-module=integreat_cms.backend.settings
+django-settings-module=integreat_cms.core.settings
 
 # Pickle collected data for later comparisons.
 persistent=yes
