+++ example-configs/apache2-integreat-vhost.conf
@@ -26,7 +26,7 @@
         # Configure the number of Django processes
         WSGIDaemonProcess example.com processes=8 threads=1 python-home=/opt/integreat-cms/.venv/ python-path=/opt/integreat-cms
         WSGIProcessGroup example.com
-        WSGIScriptAlias / /opt/integreat-cms/integreat_cms/backend/wsgi.py
+        WSGIScriptAlias / /opt/integreat-cms/integreat_cms/core/wsgi.py
 
         # Integreat CMS Settings
         SetEnv DJANGO_SECRET_KEY SECRET
