+++ integreat_cms/cms/views/pages/page_actions.py
@@ -281,7 +281,7 @@ def expand_page_translation_id(request, short_url_id):
     :param short_url_id: The id of the requested page
     :type short_url_id: int
 
-    :return: A redirection to :class:`~integreat_cms.backend.settings.WEBAPP_URL`
+    :return: A redirection to :class:`~integreat_cms.core.settings.WEBAPP_URL`
     :rtype: ~django.http.HttpResponseRedirect
     """
