+++ integreat_cms/core/sphinx_settings.py
