+++ integreat_cms/integreat-cms-cli
@@ -7,7 +7,7 @@ import sys
 
 def main():
     """" Application entry point """
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "integreat_cms.backend.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "integreat_cms.core.settings")
     running_tests = False
     for arg in sys.argv:
         if arg == "--set=COVERAGE":
